Separate monitoring data from components. Fixes #518
The repository providing the monitoring data is required by components.
But coupling the repository to application makes it hard to reuse
components without repository.

The repository is still required for components but can be separated
from monitoring statement creation. So you shall be able to stop
monitoring components but use the monitoring components e.g. as a
performance monitor.
