--- jamon-parent/jamon/src/main/java/org/wicketstuff/jamon/component/JamonRepositoryKey.java
@@ -1,11 +1,0 @@
-package org.wicketstuff.jamon.component;
-
-import org.apache.wicket.MetaDataKey;
-import org.wicketstuff.jamon.monitor.JamonRepository;
-
-public class JamonRepositoryKey extends MetaDataKey<JamonRepository>
-{
-
-	private static final long serialVersionUID = 1L;
-	public static final JamonRepositoryKey KEY = new JamonRepositoryKey();
-}
