+++ jamon-parent/jamon/src/test/java/org/wicketstuff/jamon/webapp/JamonWebApplication.java
@@ -19,7 +19,6 @@ package org.wicketstuff.jamon.webapp;
 import org.apache.wicket.markup.html.WebPage;
 import org.apache.wicket.protocol.http.WebApplication;
 import org.wicketstuff.jamon.component.JamonTestUtil;
-import org.wicketstuff.jamon.request.cycle.JamonAwareRequestCycleListener;
 
 /**
  * WebApplication so Jamon Monitoring can be tested.
@@ -43,13 +42,5 @@ public class JamonWebApplication extends WebApplication
 	public Class<? extends WebPage> getHomePage()
 	{
 		return HomePage.class;
-	}
-
-	@Override
-	protected void init()
-	{
-		super.init();
-
-		getRequestCycleListeners().add(new JamonAwareRequestCycleListener(this, true));
 	}
 }
