+++ jamon-parent/jamon/src/test/resources/META-INF/services/org.apache.wicket.IInitializer
@@ -1,0 +1,1 @@
+org.wicketstuff.jamon.application.JamonInitializer
